chore: rename folder from mctab to mc
